refactor: 이미지 업로드를 담당하는 클래스로 분리
